+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [Fix] Show deprecation warnings for legacy plot API (#513)
 * [Fix] Fix error when trying to access previously non-existing file (#840)
 * [Fix] Testing with latest DuckDB version (#498)
+* [Fix] Remove duplicate integration tests (#827)
 
 ## 0.10.1 (2023-08-30)